+++ src/eca/server.clj
@@ -62,6 +62,12 @@
 (defmethod lsp.server/receive-request "chat/delete" [_ components params]
   (handlers/chat-delete (with-config components) params))
 
+(defmethod lsp.server/receive-notification "mcp/stopServer" [_ components params]
+  (handlers/mcp-stop-server (with-config components) params))
+
+(defmethod lsp.server/receive-notification "mcp/startServer" [_ components params]
+  (handlers/mcp-start-server (with-config components) params))
+
 (defn ^:private monitor-server-logs [log-ch]
   ;; NOTE: if this were moved to `initialize`, after timbre has been configured,
   ;; the server's startup logs and traces would appear in the regular log file